requirements.txt
requests
fake-useragent
selenium
python-dotenv
pycryptodome